+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.1 - [DD/MM/YYYY]
 
 - Fix `animate` parameter on differents methods (didn't work as expected)
+- Add fields (`canScroll`, `canForward`, `canBackward`)
 
 ## 0.2.0 - 24/12/2021